% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_picture.R
\name{get_picture}
\alias{get_picture}
\title{Get the artworks pictures within a classification}
\usage{
get_picture(classification = "Prints", size = 10, numberp = 6)
}
\arguments{
\item{classification}{An art classification according to Harvard Art Museum website. The full list of available
classification is: 'Prints', 'Albums', 'Amulets', 'Armor', 'Boxes', 'Calligraphy', 'Cameos', 'Fragments',
'Furnishings', 'Gems', 'Inscriptions', 'Jewelry', 'Mirrors', 'Mosaics', 'Paintings', 'Photographs', 'Plaques',
'Rubbings', 'Sculpture', 'Seals', 'Tablets', 'Tokens', 'Vessels'. Default is set to be 'Prints'.}

\item{size}{Size specifies how many pieces of artwork the user would like to get. Default is set to be 10.}

\item{numberp}{Number of pictures that user want to download from the website.Default is set to be 7.}
}
\value{
pictures that are under the specified classification (classification1 input) with the number equals to
    specificed numberp downloaded on the user's local folder.
}
\description{
Get the artworks pictures within a classification
}
\examples{
get_picture(size=10, numberp = 2, classification = 'Vessels')
get_picture(size=5, numberp = 2, classification = 'Prints')
}
