% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_artwork.R, R/get_gallery_info.R,
%   R/get_token.R, R/most_object.R
\name{get_token}
\alias{get_token}
\title{Get a token and save it as a global variable for future use.}
\usage{
get_token()

get_token()

get_token()

get_token()
}
\value{
It will return the token that user input

It will return the token that user input

It will return the token that user input

It will return the token that user input
}
\description{
Get a token and save it as a global variable for future use.

Get a token and save it as a global variable for future use.

Get a token and save it as a global variable for future use.

Get a token and save it as a global variable for future use.
}
