% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_artwork.R, R/get_gallery_info.R,
%   R/isclassification.R, R/most_object.R
\name{is_classification}
\alias{is_classification}
\title{Check if the classification is a valid input.}
\usage{
is_classification(class)

is_classification(class)

is_classification(class)

is_classification(class)
}
\arguments{
\item{class}{A classification of Harvard Art Museum website.}
}
\value{
If this is a valid classification input for this package, it will return TRUE, otherwise, return FALSE

If this is a valid classification input for this package, it will return TRUE, otherwise, return FALSE

If this is a valid classification input for this package, it will return TRUE, otherwise, return FALSE

If this is a valid classification input for this package, it will return TRUE, otherwise, return FALSE
}
\description{
Check if the classification is a valid input.

Check if the classification is a valid input.

Check if the classification is a valid input.

Check if the classification is a valid input.
}
\examples{
is_classification(class = 'Prints')
is_classification(class = 'Music')
is_classification(class = 'Prints')
is_classification(class = 'Music')
is_classification(class = 'Prints')
is_classification(class = 'Music')
is_classification(class = 'Prints')
is_classification(class = 'Music')
}
