% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/most_object.R
\name{most_object}
\alias{most_object}
\title{Which gallery has the most artwork object on a specified floor?}
\usage{
most_object(floor = 2)
}
\arguments{
\item{floor}{The level of Harvard Art Museum. Valid floor from 1 to 3. Floor 4 is for office use. Default to be 2.}
}
\value{
a gallery name which has the most objects under a specific floor.
}
\description{
Which gallery has the most artwork object on a specified floor?
}
\examples{
most_object(floor = 3)
most_object(floor = 2)
}
