% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_token.R
\name{get_token}
\alias{get_token}
\title{Get a token and save it as a global variable for future use.}
\usage{
get_token()
}
\value{
It will return the token that user input
}
\description{
Get a token and save it as a global variable for future use.
}
